Remove ActivityPackSequenceGetter
