+++ services/QuillLMS/spec/services/independent_practice_activity_packs_assigner_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe IndependentPracticeActivityPacksAssigner do
       assigning_all_recommendations: true,
       classroom_id: classroom.id,
       diagnostic_activity_id: create(:diagnostic_activity).id,
-      release_method: ActivityPackSequence::STAGGERED_RELEASE,
+      release_method: release_method,
       selections: selections,
       user: user
     )
@@ -28,6 +28,16 @@ RSpec.describe IndependentPracticeActivityPacksAssigner do
     ActionController::Parameters.new(
       id: create(:unit_template).id,
       classrooms: [{ student_ids: student_ids2, order: 1 }]
+    )
+  end
+
+  let(:release_method) { ActivityPackSequence::STAGGERED_RELEASE }
+
+  let(:activity_pack_sequence) do
+    create(:activity_pack_sequence,
+      classroom_id: classroom_id,
+      diagnostic_activity_id: diagnostic_activity_id,
+      release_method: ActivityPackSequence::STAGGERED_RELEASE
     )
   end
 
@@ -66,6 +76,38 @@ RSpec.describe IndependentPracticeActivityPacksAssigner do
     it 'assigns two recommendations' do
       expect(AssignRecommendationsWorker).to receive(:perform_async).twice
       subject
+    end
+  end
+
+  context 'release method is staggered release' do
+    let(:release_method) { ActivityPackSequence::STAGGERED_RELEASE }
+
+    context 'Activity Pack Sequence does not exist' do
+      it { expect { subject }.to change(ActivityPackSequenceActivityPack, :count).from(0).to(1) }
+    end
+
+    context 'Activity Pack Sequence already exists' do
+      before { activity_pack_sequence }
+
+      it { expect { subject }.not_to change(ActivityPackSequenceActivityPack, :count).from(1) }
+    end
+  end
+
+  context 'release method is not staggered release' do
+    let(:release_method) { 'immediate' }
+
+    it { expect(subject).to eq nil }
+
+    context 'Activity Pack Sequence already exists' do
+      before { activity_pack_sequence }
+
+      it { expect { subject }.to change(ActivityPackSequence, :count).from(1).to(0) }
+
+      context 'Activity Pack Sequence Activity Packs exist' do
+        before { create(:activity_pack_sequence_activity_pack, activity_pack_sequence: activity_pack_sequence) }
+
+        it { expect { subject }.to change(ActivityPackSequenceActivityPack, :count).from(1).to(0) }
+      end
     end
   end
 end
